Провел корректировки для тестовых точек, теперь они работают кроме отрицательных напряжений, графики очищаются после завершения тестирования, исправлены не точности, коммит рабочий

+++ PC/mainwindow.h
@@ -44,7 +44,7 @@ private slots:
     void handleParsedPacket();
     void startTesting();
     void result(uint8_t* packet);
-    void handleCaseCommon(uint16_t sample, int ratio, const QString& labelText);
+    void handleCaseCommon(double sample, double ratio, const QString& labelText);
     void setupPort();
     void peltie(uint16_t sample, uint16_t bit);
     void setupConnections();
